+++ banking_api_ui/src/components/education/EducationPanelsHost.js
@@ -13,6 +13,7 @@ import StepUpPanel from './StepUpPanel';
 import PingOneAuthorizePanel from './PingOneAuthorizePanel';
 import CimdPanel from './CimdPanel';
 import HumanInLoopPanel from './HumanInLoopPanel';
+import BestPracticesPanel from './BestPracticesPanel';
 
 /**
  * Mounts all education drawers/modals; visibility controlled by EducationUIContext.
@@ -33,6 +34,7 @@ export default function EducationPanelsHost() {
       <PingOneAuthorizePanel isOpen={panel === EDU.PINGONE_AUTHORIZE} onClose={close} initialTabId={tab} />
       <CimdPanel isOpen={panel === EDU.CIMD} onClose={close} initialTabId={tab} />
       <HumanInLoopPanel isOpen={panel === EDU.HUMAN_IN_LOOP} onClose={close} initialTabId={tab} />
+      <BestPracticesPanel isOpen={panel === EDU.BEST_PRACTICES} onClose={close} initialTabId={tab} />
     </>
   );
 }